Added Authentication to CRUD function with Error handling methods

+++ sample_service/queries/cardio.py
@@ -1,6 +1,9 @@
-from models import CardioExerciseIn, CardioExerciseOut, CardioWorkoutIn, CardioWorkoutOut
+from models import CardioExerciseIn, CardioExerciseOut, Error, CardioWorkoutIn, CardioWorkoutOut
 from .client import Queries
+from fastapi import HTTPException, status
 from bson.objectid import ObjectId
+from typing import Union
+
 # from typing import List
 
 
@@ -8,44 +11,53 @@ class CardioQueries(Queries):
     COLLECTION = "cardio"
 
 
-    def create(self, info: CardioExerciseIn) -> CardioExerciseOut:
+    def create(self, info: CardioExerciseIn, user_id: str) -> CardioExerciseOut:
         props = info.dict()
+        props["user_id"] = user_id # Saved
         self.collection.insert_one(props)
-        props["id"] = str(props["_id"])
+        props["id"] = str(props["_id"]) # Not Saved
         return CardioExerciseOut(**props)
 
 
-    def get_all(self) -> list[CardioExerciseOut]:
+    def get_all(self, user_id: str) -> list[CardioExerciseOut]:
         CardioExercises = []
-        for exercise in self.collection.find():
+        for exercise in self.collection.find({"user_id": user_id}):
             exercise["id"] = str(exercise["_id"])
             CardioExercises.append(CardioExerciseOut(**exercise))
         return CardioExercises
-        # try:
-        # except:
-        #     return "Error Creating Cardio Workout"
 
 
-    def get_one(self, id: str) -> CardioExerciseOut :
-        result = self.collection.find_one({"_id": ObjectId(id)})
-        result["id"] = str(result["_id"])
-        return result
+    def get_one(self, id: str, user_id: str) -> Union[CardioExerciseOut, Error]:
+        try:
+            result = self.collection.find_one({"_id": ObjectId(id), "user_id": user_id})
+            result["id"] = str(result["_id"])
+            return result
+        except Exception as e:
+            raise HTTPException(
+                status_code = status.HTTP_400_BAD_REQUEST,
+                detail = "Could not find id"
+            )
 
 
-    def delete_one(self, id: str) -> bool:
-        result = self.collection.delete_one({"_id": ObjectId(id)})
+    def update(self, id: str, exercise: CardioExerciseIn, user_id: str) -> Union[CardioExerciseOut, Error]:
+        try:
+            result = self.collection.find_one({"_id": ObjectId(id), "user_id": user_id})
+            temp = exercise.dict()
+            result["id"] = str(result["_id"])
+            result["name"] = temp["name"]
+            newvalues = {"$set": result}
+            self.collection.update_one({
+                "_id": ObjectId(id),
+                }, newvalues)
+            result = self.collection.find_one({"_id": ObjectId(id), "user_id": user_id})
+            return result
+        except Exception as e:
+            raise HTTPException(
+                status_code = status.HTTP_400_BAD_REQUEST,
+                detail = "Could not Update.  Confirm response model and user id."
+            )
+
+
+    def delete_one(self, id: str, user_id: str) -> bool:
+        result = self.collection.delete_one({"_id": ObjectId(id), "user_id": user_id})
         return result.deleted_count == 1
-
-
-    def update(self, id: str, exercise: CardioExerciseIn) -> bool:
-        result = self.collection.find_one({"_id": ObjectId(id)})
-        temp = exercise.dict()
-        result["id"] = str(result["_id"])
-        result["name"] = temp["name"]
-        newvalues = {"$set": result}
-
-        output = (self.collection.update_one({
-            "_id": ObjectId(id),
-            }, newvalues))
-
-        return output.matched_count == 1
